+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/support/netty/Jt808TerminalHeatBeatHandler.java
@@ -5,7 +5,6 @@ import io.github.hylexus.jt.jt808.spec.session.Jt808SessionManager;
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
-import io.netty.handler.timeout.IdleState;
 import io.netty.handler.timeout.IdleStateEvent;
 import lombok.extern.slf4j.Slf4j;
 
@@ -20,12 +19,15 @@ import static io.github.hylexus.jt.jt808.spec.session.DefaultSessionCloseReason.
 @BuiltinComponent
 public class Jt808TerminalHeatBeatHandler extends ChannelInboundHandlerAdapter {
 
-    private final Jt808SessionManager sessionManager;
+    protected final Jt808SessionManager sessionManager;
 
     public Jt808TerminalHeatBeatHandler(Jt808SessionManager sessionManager) {
         this.sessionManager = sessionManager;
     }
 
+    /**
+     * @see <a href="https://github.com/hylexus/jt-framework/issues/66">https://github.com/hylexus/jt-framework/issues/66</a>
+     */
     @Override
     public void userEventTriggered(ChannelHandlerContext ctx, Object evt) throws Exception {
         if (!(evt instanceof IdleStateEvent)) {
@@ -33,9 +35,22 @@ public class Jt808TerminalHeatBeatHandler extends ChannelInboundHandlerAdapter {
             return;
         }
 
-        if (((IdleStateEvent) evt).state() == IdleState.READER_IDLE) {
-            log.debug("disconnected idle connection");
+        try {
+            log.debug("disconnected idle connection, reason: {}", ((IdleStateEvent) evt).state());
             sessionManager.removeBySessionIdAndClose(sessionManager.generateSessionId(ctx.channel()), IDLE_TIMEOUT);
+        } catch (Exception e) {
+            log.error(e.getMessage(), e);
         }
+
+        // 之所以有下面这个再次关闭 channel 的代码, 是因为上面的 sessionManager.removeBySessionIdAndClose() 可能无法真正关闭对应的 channel
+        // 比如: 即便是客户端建立了连接，但是客户端没有发送任何消息的情况下, sessionManager 里是不会有对应的session的；
+        // 所以通过 session.channel().close() 是关不掉的
+        // @see https://github.com/hylexus/jt-framework/issues/66
+        try {
+            ctx.channel().close();
+        } catch (Exception ignored) {
+            // ignored
+        }
+
     }
 }