+++ samples/jt-808-server-sample-annotation/src/main/java/io/github/hylexus/jt808/samples/annotation/model/vo/TerminalVo.java
@@ -1,6 +1,7 @@
 package io.github.hylexus.jt808.samples.annotation.model.vo;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
@@ -15,4 +16,7 @@ public class TerminalVo {
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private Date lastCommunicationTime;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private String latestGeo;
 }